requirements.txt
OpenAI
gradio
python-dotenv